+++ pl_bolts/models/detection/retinanet/__init__.py
@@ -1,4 +1,4 @@
 from pl_bolts.models.detection.retinanet.backbones import create_retinanet_backbone
-from pl_bolts.models.detection.retinanet.retainanet_module import RetinaNet
+from pl_bolts.models.detection.retinanet.retinanet_module import RetinaNet
 
 __all__ = ["create_retinanet_backbone", "RetinaNet"]
